+++ packages/db/prisma/migrations/20260804200542_add_schedule_recent/migration.sql
@@ -1,0 +1,2 @@
+-- AlterTable
+ALTER TABLE "channel" ADD COLUMN     "scheduleRecent" JSONB;